+++ src/app/(route)/my/profile/_components/ProfileInputArea.tsx
@@ -1,12 +1,12 @@
 import { TProps } from "@/types/component/props";
 
-export function ProfileInputArea({
+export default function ProfileInputArea({
   label,
   children,
-}: { label: string } & TProps) {
+}: { label?: string } & TProps) {
   return (
     <div className="flex flex-col gap-2 justify-start">
-      <p className="text-H4">{label}</p>
+      {label && <p className="text-H4">{label}</p>}
       {children}
     </div>
   );
